+++ parsec_android/lib/parsec_android.dart
@@ -17,9 +17,7 @@ class ParsecAndroid extends ParsecPlatform {
 
   @override
   Future<dynamic> nativeEval(String equation) {
-    return _channel.invokeMethod(
-      'nativeEval',
-      {'equation': equation},
-    ).then((result) => result);
+    return _channel.invokeMethod('nativeEval', {'equation': equation}).then(
+        (result) => parseNativeEvalResult(result));
   }
 }