+++ parsec_linux/lib/parsec_linux.dart
@@ -1,7 +1,5 @@
-import 'dart:convert';
 import 'package:flutter/services.dart';
 import 'package:parsec_platform_interface/parsec_platform_interface.dart';
-import 'package:parsec_platform_interface/parsec_eval_exception.dart';
 
 const MethodChannel _channel = MethodChannel('parsec_linux');
 
@@ -14,27 +12,5 @@ class ParsecLinux extends ParsecPlatform {
   Future<dynamic> nativeEval(String equation) {
     return _channel.invokeMethod('nativeEval', {'equation': equation}).then(
         (result) => parseNativeEvalResult(result));
-  }
-
-  dynamic parseNativeEvalResult(String jsonString) {
-    var jsonData = jsonDecode(jsonString);
-    var val = jsonData['val'];
-    var type = jsonData['type'];
-    var error = jsonData['error'];
-
-    if (error != null) {
-      throw ParsecEvalException(error);
-    }
-
-    switch (type) {
-      case 'i':
-        return int.parse(val);
-      case 'f':
-        return double.parse(val);
-      case 'b':
-        return val == 'true';
-      default:
-        return val;
-    }
   }
 }